requirements.txt
playwright
icalendar
google-api-python-client
google-auth
google-auth-oauthlib
